+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## Unreleased
+## v1.0.0 (2026-01-19)
+
+### BREAKING CHANGE
+
+- Task.run method had one of its kwargs removed
+
+### Fix
+
+- **Task-can-no-longer-pass-logger-to-its-function.-Changed-some-examples,-documentations-and-better-type-hints**: Task.run method no longer can pass logger to its function as kwarg
 
 ## v0.1.0 (2026-01-18)